\name{newSOCKnode}
\alias{newSOCKnode}
\title{newSOCKnode}
\description{adapted from snow 0.3-3, modified to work with pahts that include white spaces}
\usage{
newSOCKnode(machine = "localhost",
...,
options = defaultClusterOptions)
}
\arguments{
  \item{machine}{
}
  \item{\dots}{
}
  \item{options}{
}
}



\author{Thomas Wutzler}





