\name{twResultFilename}
\alias{twResultFilename}
\title{twResultFilename}
\description{Construct a modified filename from paramFilename.}
\usage{
twResultFilename(paramFilename,
iproc = NULL,
prefix = "res_",
ext = ".RData")
}
\arguments{
  \item{paramFilename}{filename to modify}
  \item{iproc}{a number to append to basename before .RData (if NULL nothing is appended)}
  \item{prefix}{prefix to prepend to basename}
  \item{ext}{file extension}
}



\author{Thomas Wutzler}





