\name{twSnowfall-package}
\alias{twSnowfall-package}
\title{tw snowfall extension}
\description{snowfall extension for 1) extended load balancing and 2) dependencies of arguments to remote function on former results, and 3) Simplifying debugging of remote errors.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}

\keyword{ package }


